vdpa/virtio: add VM UUID devargs
This commit adds a new devargs named vm_uuid to virtio vdpa driver.
This UUID is a unique ID of VM, most likely the UUID of libvirt VM.

To specify a UUID, user could use '-a 0000:af:00.1 -u $UUID' to let driver
know the VM UUID of one virtio VF. This parameter is useful for later
optimization of DMA mapping.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>