app/vfe-vdpa: add HA restart module
This commit adds the HA restart module in vDPA application. The restart
flow supports working with HA service or not. When HA service is there,
it will query information from HA service and try to restore devices.
When HA service is not alive, it will skip the restore when calling
virtio_ha_pf_list_query().

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>