common/virtio_mi: support HA store and remove context
This commit adds support of device context store and remove in
virtio PF driver. When device context is set from application,
the driver will restore from the context and will not store context.
Otherwise, the driver will initialize everything and store the
context. The store and restore happens in driver probe and remove
happens in driver remove.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>