vdpa/virtio: add connection timeout support
When HA service is alive, it holds the DMA mapping of the hugepages.
There is one corner case that QEMU could fail to restart:

1. vDPA application crashes
2. QEMU crashes
3. vDPA application restarts and restore all device context
4. QEMU tries to restart but failed due to not enough hugepages (HA
service holding the hugepages).

To resolve this potential issue, this commit adds a connection timeout
mechanism in vhost library and vdpa driver, which cleans up the DMA
mapping of the hugepages when vDPA application can't connect to QEMU
for 3 seconds.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>