common/virtio_ha: add cache layer and sync mechanism
When vDPA application store or remove context to HA service, it could
happen that HA service is not alive. When HA service becomes alive
after store or remove finishes, we need a mechanism to sync the context
to HA service. In this commit, a cache layer is introduced on the IPC
client side which saves all context information in HA library. Then a sync
mechanism is introduced to sync to context when HA library detects that
HA service is alive again.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>