vhost: delete cross numa reallocation mechanism
In vhost library, there exists a mechanism to reallocate vq resources
when numa node of VM memory is different from that of current vq memory.
The reallocation could use much time (more than 200ms as measured). This
leads to vDPA application restart very slowly. So this commit delete it.

The impact should be minimal as it will not impact date path performance
when using vDPA. It could increase the dirty page logging time as the
DMA of dirty page logging could be cross-numa. So it could increase the
migration time but not migration downtime, which is more important.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>