vdpa/virtio: replace HPA with QEMU process ID
Before this commit, we use HPA for checking if old and new memory region
is the same or not. This is for a corner case that when vhostd restart,
qemu also restart, then qemu could send different memory region with same
info (QEMU_VA, GPA, SIZE). Previously we use HPA to handle this case,
but the side effect is we need to use MAP_POPULATE flag for mmap call,
which results in more time used in mmap. In real environment, the time
could be several seconds when mmap hundreds of GB memory.

This commit removes the usage of HPA and MAP_POPULATE flag, but use QEMU
process id to handle the corner case.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>