common/virtio_ha: add lock for IPC send
sendmsg() API is thread-safe but could be not atomic if the send buffer
is larger than PIPE_BUF (4096 bytes). It has small possibility that the
send buffer could be very large. To ensure safely, this commit adds the
send lock for IPC socket. For receive, there will be no multiple thread
calling recvmsg(), so no lock is needed.

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>