vdpa/virtio: remove HPA for memory table
Previously, we use HPA to handle corner case when vhostd and qemu crash
at the same time, but to get HPA, we add a mmap flag MAP_POPULATE, which
makes the mmap time very long. This commit removes the usage of HPA and
let vhostd-ha to check vhost fd connection status to know if QEMU
crashes or not. If QEMU crashed, vhostd cleanup the memory table on
restart. Meanwhile, we also handle the case of device unplug, which will
also leads to socket disconnect.

RM: 3958103

Signed-off-by: Chenbo Xia <chenbox@nvidia.com>